+++ profileservice/configs.py
@@ -1,8 +1,17 @@
 import os
+
+from dotenv import load_dotenv
+
+# Load .env file
+load_dotenv()
+
 # this has to be have the file path to the folder for saving image files and other necessary files from rest service
 PROFILE_REST_STORAGE = os.getenv('REST_STORAGE', '/usr/src/app/rest')
 MONGO_PROFILE_URL = os.getenv('MONGO_PROFILE_URL', 'localhost:27017')
 MONGO_PII_URL = os.getenv('MONGO_PII_URL', 'localhost:27017')
+FLASK_APP = os.getenv('FLASK_APP', 'profile_rest_service')
+FLASK_ENV = os.getenv('FLASK_ENV', 'production')
+PROFILE_ENDPOINT = os.getenv('PROFILE_ENDPOINT', '/profiles')
 
 PROFILE_DB_NAME = 'profiledb'
 PII_DB_NAME = 'piidb'