+++ profileservice/restservice/utils/rest_handlers.py
@@ -19,9 +19,10 @@ def return_id(msg, idfield, id):
 
     return make_response(resp)
 
-def entry_deleted(id):
+
+def entry_deleted(id_type, id):
     message = {
-        'message': 'Object is deleted with id of : ' + id
+        id_type: id
     }
     resp = jsonify(message)
     resp.status_code = 202
@@ -30,62 +31,64 @@ def entry_deleted(id):
 
 @app.errorhandler(400)
 def bad_request(error=None):
-    message = {
-        'message': 'Bad Request: ' + request.url,
-        'reason': error
-    }
-    resp = jsonify(message)
+    if error is None:
+        error = {
+            'error': 'Bad Request: ' + request.url,
+        }
+    resp = jsonify(error)
     resp.status_code = 400
 
     return resp
 
 @app.errorhandler(403)
 def forbidden(error=None):
-    message = {
-        'message': 'Forbidden: ' + request.url,
-    }
-    resp = jsonify(message)
+    if error is None:
+        error = {
+            'error': 'Forbidden: ' + request.url,
+        }
+    resp = jsonify(error)
     resp.status_code = 403
 
     return resp
 
 @app.errorhandler(404)
 def not_found(error=None):
-    if error is not None:
-        message = {'message': error + ": " + request.url}
-    else:
-        message = {'message': 'Not Found: ' + request.url}
-    resp = jsonify(message)
+    if error is None:
+        error = {'error': 'Not Found: ' + request.url}
+    resp = jsonify(error)
     resp.status_code = 404
 
     return resp
 
 @app.errorhandler(415)
 def unsupported_media_type(error=None):
-    message = {
-        'message': 'Unsupported media type: ' + request.url,
-    }
-    resp = jsonify(message)
+    if error is None:
+        error = {
+            'error': 'Unsupported media type: ' + request.url,
+        }
+    resp = jsonify(error)
     resp.status_code = 415
 
     return resp
 
 @app.errorhandler(500)
 def internal_server_error(error=None):
-    message = {
-        'message': 'Internal Server Error: ' + request.url,
-    }
-    resp = jsonify(message)
+    if error is None:
+        error = {
+            'error': 'Internal Server Error: ' + request.url,
+        }
+    resp = jsonify(error)
     resp.status_code = 500
 
     return resp
 
 @app.errorhandler(501)
 def not_implemented(error=None):
-    message = {
-        'message': 'Not Implemented: ' + request.url,
-    }
-    resp = jsonify(message)
+    if error is None:
+        error = {
+            'error': 'Not Implemented: ' + request.url,
+        }
+    resp = jsonify(error)
     resp.status_code = 501
 
     return resp